Added initial support for runtime-generic bootstrap.
ASM macro allows embedding primops directly in the implementations of
the bootstrap functions, but that proved not to be too useful - the main
issue lies in &apply and &yield-cont which are then wrapped in another
similar call during compilation, resulting in weird behaviour. Instead,
a set of generic continuation-manipulating function is assumed to exist
- yield, continue-with* - in terms of which then the other functions are
  defined.
Known issues: &current-continuation primop is expanded during code-gen,
meaning that its lexical placement may change in case the containing
function is inlined. That results in weird continuation behaviour.
Ideally, it would be expanded during CPS.